fix: Properly organize tests into mirrored namespace, resolve issues that caused tests to fail (yay! tests helped!)
