+++ Assets/Tests/PlayMode/ClientSessionTests.cs
@@ -30,7 +30,7 @@ namespace PlayEveryWare.EpicOnlineServices.Tests.ClientTests
     using System.Collections;
     using UnityEngine;
     using UnityEngine.TestTools;
-    using JsonUtility = Utility.JsonUtility;
+    using JsonUtility = PlayEveryWare.EpicOnlineServices.Utility.JsonUtility;
 
     /// <summary>
     /// Session connection tests that test connecting to an existing session.
