+++ Assets/Tests/PlayMode/Utility/Extensions/CustomMappedEnumTestBase.cs
